Renamed listinfo into model
