+++ src/components/UserMenuButton.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
-import UserInfo from "../listdata/userinfo";
-import { AnimeStatus } from "../listdata/api";
-import { COLORS, ThemeData, BRIGHTNESES } from "../listdata/theme";
+import UserInfo from "../model/userinfo";
+import { AnimeStatus } from "../model/api";
+import { COLORS, ThemeData, BRIGHTNESES } from "../model/theme";
 
 const BRIGHTNESS_LABELS = { dark: "Dark", auto: "Auto", light: "Light" };
