+++ src/model/seriesinfo.ts
